% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GCD_pairs.R
\name{GCD_pairs}
\alias{GCD_pairs}
\title{Travel distance}
\usage{
GCD_pairs(x_lon, x_lat, y_lon, y_lat)
}
\arguments{
\item{x_lon}{lon}

\item{x_lat}{lat}

\item{y_lon}{lon}

\item{y_lat}{lat}
}
\value{
double
}
\description{
give a set of location L, it calculate the travelling distance, i.e. sum of all the distance from L(i) to L(i-1).
}
\author{
Yi Tang
}

