% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Parse_GPX.R
\name{Parse_GPX}
\alias{Parse_GPX}
\title{Parse GPX file}
\usage{
Parse_GPX(file.path)
}
\arguments{
\item{file.path}{file path to .GPX file}
}
\value{
list
}
\description{
main functions to parse GPX file, see package description for more details
}
\author{
Yi Tang
}

