% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Parse_GPX_all.R
\name{Parse_GPX_all}
\alias{Parse_GPX_all}
\title{Parse GPX file}
\usage{
Parse_GPX_all(data.dir = NULL, app = "Strava", add.city = TRUE)
}
\arguments{
\item{data.dir}{a folder that contains GPX files}

\item{app}{RunKeeper, Strave or others?}

\item{add.city}{add city and country}
}
\value{
a list
}
\description{
A warpper function for Parse_GPX.
}
\author{
Yi Tang
}

