% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RExercise.R
\docType{package}
\name{RExercise}
\alias{RExercise}
\alias{RExercise-package}
\title{RExercise}
\description{
This package contains extra themes, scales, and geoms, and
functions for and related to \pkg{ggplot2}.
}
\details{
In addition to the help pages, see the README page on
\href{https://github.com/jrnold/ggthemes}{github} for examples.
}

